% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyISO52010.R
\name{add_dayOfYear_hourOfDay}
\alias{add_dayOfYear_hourOfDay}
\title{add helper variables}
\usage{
add_dayOfYear_hourOfDay(.df, col_timestamp = "timestamp")
}
\arguments{
\item{.df}{input data frame}

\item{col_timestamp}{name of the column holding timestamp information, defaults to "timestamp"}
}
\value{
.df with columns n_day and n_hour added. Timestamp "2014-01-02 00:00" results in n_day = 1 & n_hour = 24.
n_day is the day of the year, an integer between 1-366
n_hour is the hour of the day, an decimal between 0.01667 - 24
}
\description{
adds columns n_day and n_hour to .df
}
