% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_ISO52010}
\alias{rcpp_ISO52010}
\title{ISO 52010-1:2017 solar irradiance calculation}
\usage{
rcpp_ISO52010(lat, lng, tz, t_shift, surfaceAzimuths, surfaceTilts, n_hour,
  n_day, G_dir, G_dif, albedo, interp_perez)
}
\arguments{
\item{lat}{Latitude, in radians}

\item{lng}{Longitude, in radians}

\item{tz}{Time zone of the irradiance data, in hours. E.g. +1 for central European 
(UTC+1) time zones. Use 0, if data is recorded in UTC time.}

\item{t_shift}{Decimal hours to shift timestamps with. E.g. timestamp 12:00 usually refer to the
solar irradaince during time interval 11:00 - 12:00, then use t_shift = 0.5. 
If the timestamp 12:00 refer to interval 11:45 - 12:15, use t_shift = 0.0. 
If the timestamp 12:00 refer to interval 11:45 - 12:00, use t_shift = 0.125.}

\item{surfaceAzimuths}{Vector of surface azimuths, in radians}

\item{surfaceTilts}{Vector of surface tilts, in radians. Need to be same lenght as surfaceAzimuths}

\item{n_hour}{Vector representing the hour of the day for every time step (double 0-24)}

\item{n_day}{Vector representing the day of the year for every time step (integer 1-366)}

\item{G_dir}{Vector of direct normal irradiance, W/m2}

\item{G_dif}{Vector of diffuse horizontal irradiance, W/m2}

\item{albedo}{Vector of solar reflectivity of the ground, 0.2 can be used if actual conditions are unknown.}

\item{interp_perez}{Set to 0 to interpolate from the Parez table (clearness index    and    brightness coefficients). 
Set to 1 to use binned values, this follow the procedure of the ISO 52010-1:2017 standard.}
}
\value{
A matrix where first column holds the solar altitude for each time step,
from second column calculated total direct irradiances (I_tot_dir) for each surface are returned, 
and from column position 1 + "number of surfaces" total diffuse irradiances are returned
}
\description{
Calculate solar irradiance on a surface with arbitrary orientation and tilt
}
