% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_ISO52010_angles}
\alias{rcpp_ISO52010_angles}
\title{ISO 52010-1:2017 solar altitude and azimuth}
\usage{
rcpp_ISO52010_angles(lat, lng, tz, t_shift, n_hour, n_day)
}
\arguments{
\item{lat}{Latitude, in radians}

\item{lng}{Longitude, in radians}

\item{tz}{Time zone of the irradiance data, in hours. E.g. +1 for central European 
(UTC+1) time zones. Use 0, if data is recorded in UTC time.}

\item{t_shift}{Decimal hours to shift timestamps with. E.g. timestamp 12:00 usually refer to the
solar irradaince during time interval 11:00 - 12:00, then use t_shift = 0.5. 
If the timestamp 12:00 refer to interval 11:45 - 12:15, use t_shift = 0.0. 
If the timestamp 12:00 refer to interval 11:45 - 12:00, use t_shift = 0.125.}

\item{n_hour}{Vector representing the hour of the day for every time step (double 0-24)}

\item{n_day}{Vector representing the day of the year for every time step (integer 1-366)}
}
\value{
A matrix where first column holds the solar altitude for each time step and second column holds solar azimuth angle
}
\description{
Calculate solar altitude and azimuth
}
