% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyISO52010.R
\name{tidyISO52010_angles}
\alias{tidyISO52010_angles}
\title{Tidy version of the ISO 52010-1:2017 solar altitude and azimuth calculation}
\usage{
tidyISO52010_angles(
  .df,
  lat,
  lng,
  tz,
  t_shift = NULL,
  col_timestamp = "timestamp"
)
}
\arguments{
\item{.df}{Data frame that holds the solar irradiance data}

\item{lat}{Latitude, in decimal degrees}

\item{lng}{Longitude, in decimal degrees}

\item{tz}{Time zone of the irradiance data, in hours. E.g. +1 for central European
(UTC+1) time zones. Use 0, if data is recorded in UTC time.}

\item{t_shift}{Decimal hours to shift timestamps with. E.g. timestamp 12:00 usually refer to the
solar irradaince during time interval 11:00 - 12:00, then use t_shift = 0.5. 
If the timestamp 12:00 refer to interval 11:45 - 12:15, use t_shift = 0.0. 
If the timestamp 12:00 refer to interval 11:45 - 12:00, use t_shift = 0.125.
If t_shift is NULL, it is estimated as (n_hour[2] - n_hour[1])/2.}
}
\description{
adds columns n_day and n_hour to .df
}
